Correct types for headerfs: seek distances should be unsigned integer 64 heights (header or filter) should be unsigned integer 32
The changes eliminate conversions.

+++ headerfs/file.go
@@ -19,7 +19,7 @@ func (h *headerStore) appendRaw(header []byte) error {
 // readRaw reads a raw header from disk from a particular seek distance. The
 // amount of bytes read past the seek distance is determined by the specified
 // header type.
-func (h *headerStore) readRaw(seekDist int64) ([]byte, error) {
+func (h *headerStore) readRaw(seekDist uint64) ([]byte, error) {
 	var headerSize uint32
 
 	// Based on the defined header type, we'll determine the number of
@@ -39,7 +39,7 @@ func (h *headerStore) readRaw(seekDist int64) ([]byte, error) {
 	// for that number of bytes, and read directly from the file into the
 	// buffer.
 	rawHeader := make([]byte, headerSize)
-	if _, err := h.file.ReadAt(rawHeader, seekDist); err != nil {
+	if _, err := h.file.ReadAt(rawHeader, int64(seekDist)); err != nil {
 		return nil, err
 	}
 
@@ -48,10 +48,10 @@ func (h *headerStore) readRaw(seekDist int64) ([]byte, error) {
 
 // readHeader reads a full block header from the flat-file. The header read is
 // determined by the hight value.
-func (h *BlockHeaderStore) readHeader(height int64) (*wire.BlockHeader, error) {
+func (h *BlockHeaderStore) readHeader(height uint32) (*wire.BlockHeader, error) {
 	// Each header is 80 bytes, so using this information, we'll seek a
 	// distance to cover that height based on the size of block headers.
-	seekDistance := height * 80
+	seekDistance := uint64(height) * 80
 
 	// With the distance calculated, we'll raw a raw header start from that
 	// offset.
@@ -72,8 +72,8 @@ func (h *BlockHeaderStore) readHeader(height int64) (*wire.BlockHeader, error) {
 
 // readHeader reads a single filter header at the specified height from the
 // flat files on disk.
-func (f *FilterHeaderStore) readHeader(height int64) (*chainhash.Hash, error) {
-	seekDistance := height * 32
+func (f *FilterHeaderStore) readHeader(height uint32) (*chainhash.Hash, error) {
+	seekDistance := uint64(height) * 32
 
 	rawHeader, err := f.readRaw(seekDistance)
 	if err != nil {
